+++ core/src/models/drink.rs
@@ -18,12 +18,22 @@ pub struct NewDrink<'a> {
     pub colour: &'a str,
 }
 
-#[derive(Clone, Debug, Queryable)]
+use diesel::sql_types::Integer;
+use diesel::sql_types::BigInt;
+use diesel::sql_types::Text;
+use diesel::sql_types::Bool;
+
+#[derive(Clone, Debug, QueryableByName)]
 pub struct DrinkWithCount {
+    #[sql_type = "Integer"]
     pub id: i32,
+    #[sql_type = "Text"]
     pub name: String,
     /// Calculated property which comes from the `core::models::drink_drank::DrankDrank`s of this `Drink`.
-    pub count: i32,
+    #[sql_type = "BigInt"]
+    pub count: i64,
+    #[sql_type = "Text"]
     pub colour: String,
+    #[sql_type = "Bool"]
     pub deleted: bool,
 }
